refactor(steerer): extract DriftObserver bucket 3a — observability cluster
Wave C bucket 3a of 3. Extracts the drift-event observability and classification surface into goldfive.drift_observer.DriftObserver (935 LOC). DefaultSteerer keeps the public surface via 18 thin shims; composition pattern with router back-reference matches buckets 1 and 2.

Scope (this PR):
- _emit_drift_detected (the canonical DriftDetected emit path)
- _stamp_drift_lifecycle + _drift_lifecycle_pb_value (#271 lifecycle stamping)
- _is_terminal_drift + _TERMINAL_DRIFT_KINDS, _close_open_boundaries_for_terminal_drift
- _resolve_authored_by + _USER_AUTHORED_DRIFT_KINDS, _drift_annotation_id (source attribution)
- detect_drift (primitive classifier dispatch, #245 observed_revision_index stamping)
- _drift_from_control + _unpack_steer_context + _is_duplicate_steer + _steer_dedupe_id (#171 control→drift mapping + dedupe)
- report_new_work_discovered + report_plan_divergence (reporting-tool drift hooks)
- _truncate_trigger_input, _summarize_recent_tool_calls, _summarize_recent_reasoning, _parse_reflective_response (display + bounded summarisation)
- _is_task_progress_stalled + _USER_OR_TRAJECTORY_DRIFT_KINDS, _emit_progress_stalled_escalation, _emit_handler_exhausted_escalation (#271 structural escalation predicates)

Honest scope-stop: the full bucket-3 brief covered ~3000 LOC across 80 items (the largest and most fragile cut). This PR ships the observability sub-cluster (low-risk, byte-equivalent moves of utility + emit-path helpers, ~700 LOC moved). Follow-up PRs cover:
- bucket 3b: observe/observe_reasoning, background judge plumbing, reflective + GOAL_DRIFT periodic checks (~500 LOC)
- bucket 3c: _handle_drift (753 lines), the intervention ladder, ladder dispatch, adapter cancel tagging, late-drift gate, request_invocation_cancel, _cancel_inflight_for_revision, _promote_drift_to_steer (351 lines), _apply_user_steer_state, refine-outcome bookkeeping (~1800 LOC). Audit issues #402 (dispatch-before-plan-swap) and #403 (lock window) live in this last cut and stay preserved-not-fixed; the focused diff makes them reviewable in isolation.

Contracts preserved (byte-equivalence verified by test suite):
- observation_only carve-outs (#264) — _dispatch_goldfive_pause_control still on router for this PR; PAUSE_ESCALATE carve-out unchanged
- Late-drift gate (#242, #230) — _is_late_drift_for_terminated_invocation still on router; reads _resolve_authored_by via shim
- Synthetic-USER_STEER suppression (#199, #183) — _apply_user_steer_state still on router for this PR
- observed_revision_index freshness (#245) — detect_drift stamps via DriftObserver; dispatch-time gate in _handle_drift on router reads the field
- Plan + Task immutability (#247): unchanged
- ControlMessage channel (#246): unchanged
- State-ownership audit: no new _KNOWN_CALLERS entries needed; no direct state[KEY]=value writes in moved code

Test suite: 2539 passed, 59 skipped; ruff clean.
LOC: steerer.py 5771 → 5229 (-542); drift_observer.py 935 LOC new.
